Add conditional
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep@38868 72102866-910b-0410-8b05-ffd578937521

+++ install.sh
@@ -1,5 +1,12 @@
 #!/bin/sh
 echo "Installing xsession..."
-install -m 0755 gnustep-wmaker /usr/local/bin
-install -m 0644 gnustep-wmaker.desktop /usr/share/xsessions
+# Check for local
+if [ -e /usr/local/bin ]; then
+    install -m 0755 gnustep-wmaker /usr/local/bin
+fi
+
+# Check for xsessions and install
+if [ -e /usr/share/xsessions ]; then
+    install -m 0644 gnustep-wmaker.desktop /usr/share/xsessions
+fi
 exit 0
